requirements.txt
opencv-python
numpy
scipy
Pillow
tqdm
ultralytics
